refactor: setter/getter error types use ErrorCode instead of i32
All public setter methods, state accessors, mapping matrix operations,
and extension parsing functions now return Result<_, ErrorCode> instead
of Result<_, i32>. Internal codec functions keep raw i32 returns;
tools/demo/backend.rs converts at the trait boundary with map_err.

+++ tests/restricted_application_parity.rs
@@ -109,7 +109,7 @@ fn set_application_restricted_modes_match_c() {
         let rust_app = Application::try_from(restricted).expect("restricted app convert");
         let rust_ret = match rust_enc.set_application(rust_app) {
             Ok(()) => 0,
-            Err(e) => e,
+            Err(e) => i32::from(e),
         };
         let c_ret = c_enc.set_application(restricted);
         assert_eq!(
@@ -129,7 +129,7 @@ fn set_application_restricted_modes_match_c() {
 
     let rust_ret = match rust_restricted.set_application(Application::Audio) {
         Ok(()) => 0,
-        Err(e) => e,
+        Err(e) => i32::from(e),
     };
     let c_ret = c_restricted.set_application(OPUS_APPLICATION_AUDIO);
     assert_eq!(
